requirements.txt
urllib3==2.3.0
requests==2.32.3
websockets==10.4
python-dotenv==1.0.1
solana==0.36.1
solders==0.23.0
discord.py==2.3.2
python-telegram-bot==20.7